#146027 self register services
